api/requirements.txt
django_recaptcha==4.0.0
fastapi==0.111.0
httpx==0.27.0
loguru==0.7.2
peewee==3.17.1
pydantic==2.7.3
Requests==2.32.3
slowapi==0.1.9
starlette==0.37.2